feat: Guard all api
